requirements.txt
pyserial == 3.5
python-pcapng == 2.1.1